+++ src/components/index.tsx
@@ -1,3 +1,4 @@
+export { List as default } from './List'
 export { Gradient } from './Gradient'
 export { InsetShadow } from './InsetShadow'
 export { Item } from './Item'
