--- src/interfaces/index.ts
@@ -1,4 +1,0 @@
-export type { Gradient as GradientProps } from './Gradient'
-export type { Item as ItemProps } from './Item'
-export type { Marker as MarkerProps } from './Marker'
-export type { Props } from './Props'
