--- src/types/Animate.ts
@@ -1,1 +1,0 @@
-export type Animate = 'scale' | 'rotate'
