+++ src/types/Gradient.ts
