--- src/types/MarkerDisplay.ts
@@ -1,1 +1,0 @@
-export type MarkerDisplay = 'adjust' | 'contrast' | string
