--- src/types/MarkerType.ts
@@ -1,1 +1,0 @@
-export type MarkerType = 'border' | 'icon' | 'fade' | 'none'
